flaskapp/requirements.txt
flask
requests
deepgram-sdk
python-dotenv
openai